2-encontro/requirements.txt
adk
python-dotenv
